Update NixOS-specific tools, part 2

+++ shell.nix
@@ -13,7 +13,9 @@ pkgs.mkShell rec {
   shellHook = ''
     export PATH=$PATH:''${CARGO_HOME:-~/.cargo}/bin
     export PATH=$PATH:''${RUSTUP_HOME:-~/.rustup}/toolchains/$RUSTC_VERSION-x86_64-unknown-linux-gnu/bin/
-    cp ${pkgs.OVMFFull.fd}/FV/OVMF.fd OVMF-pure-efi.fd
+    if [ ! -f OVMF-pure-efi.fd ]; then
+      cp ${pkgs.OVMFFull.fd}/FV/OVMF.fd OVMF-pure-efi.fd
+    fi
     '';
   
   # Needed for testing
